Fix to the TCP close detection from Fedor.
Flush the timeout queue when a connect is explicitly cancelled
but do not flush it when the link goes up.

In retry we want to go DOWN if the link is two-way (because
the other end might have deliberately hung up and can dial
back if it needs to) otherwise to RETRY (because we decide
when the link is idle).

Use "ip route replace ..." instead of "ip route add ..."

Changed pointers to web site and mailing list archive in
the README.